Format Binance earn file

Format a .csv earn history file from Binance for later ACB processing.

Usage

format_binance(data, list.prices = NULL, force = FALSE)

Arguments

data

The dataframe

list.prices

A list.prices object from which to fetch coin prices.

force

Whether to force recreating list.prices even though it already exists (e.g., if you added new coins or new dates).

Value

A data frame of exchange transactions, formatted for further processing.

Details

To get this file. Download your overall transaction report (this will include your trades, rewards, & "Referral Kickback" rewards). To get this file, connect to your Binance account on desktop, click "Wallet" (top right), "Transaction History", then in the top-right, "Generate all statements". For "Time", choose "Customized" and pick your time frame.

Warning: This does NOT process WITHDRAWALS (see the format_binance_withdrawals() function for this purpose).

Examples

# \donttest{
format_binance(data_binance)
#> Object 'list.prices' already exists. Reusing 'list.prices'. To force a fresh download, use argument 'force = TRUE'.
#>                   date currency   quantity  total.price   spot.rate transaction
#> 1  2021-05-29 17:07:20      LTC 2.53200000 5.676524e+02  224.191313         buy
#> 2  2021-05-29 17:07:20      ETH 0.19521000 5.676524e+02 2907.906384        sell
#> 3  2021-05-29 17:07:20      LTC 2.41210000 5.407719e+02  224.191313         buy
#> 4  2021-05-29 17:07:20      ETH 0.14123140 5.407719e+02 3828.977598        sell
#> 5  2021-05-29 17:07:20      LTC 1.45120000 3.253464e+02  224.191313         buy
#> 6  2021-05-29 17:07:20      ETH 0.11240000 3.253464e+02 2894.541226        sell
#> 7  2021-05-29 17:07:20      LTC 1.42100000 3.185759e+02  224.191313         buy
#> 8  2021-05-29 17:07:20      ETH 0.10512900 3.185759e+02 3030.332793        sell
#> 9  2021-05-29 17:07:20      LTC 0.30000000 6.725739e+01  224.191313         buy
#> 10 2021-05-29 17:07:20      ETH 0.00899120 6.725739e+01 7480.357904        sell
#> 11 2021-05-29 17:07:20      LTC 0.27000000 6.053165e+01  224.191313         buy
#> 12 2021-05-29 17:07:20      ETH 0.00612410 6.053165e+01 9884.171484        sell
#> 13 2021-05-29 17:07:20      LTC 0.00202500 4.539874e-01  224.191313     revenue
#> 14 2021-05-29 17:07:20      LTC 0.00127520 2.858888e-01  224.191313     revenue
#> 15 2021-05-29 17:07:20      LTC 0.00113100 2.535604e-01  224.191313     revenue
#> 16 2021-05-29 17:07:20      LTC 0.00049230 1.103694e-01  224.191313     revenue
#> 17 2021-05-29 17:07:20      LTC 0.00007000 1.569339e-02  224.191313     revenue
#> 18 2021-05-29 17:07:20      LTC 0.00005000 1.120957e-02  224.191313     revenue
#> 19 2021-05-29 18:12:55      ETH 0.44124211 1.375327e+03 3116.943792         buy
#> 20 2021-05-29 18:12:55      LTC 1.60000000 1.375327e+03  859.579285        sell
#> 21 2021-05-29 18:12:55      ETH 0.42124000 1.312981e+03 3116.943792         buy
#> 22 2021-05-29 18:12:55      LTC 1.23000000 1.312981e+03 1067.464555        sell
#> 23 2021-05-29 18:12:55      ETH 0.00021470 6.692078e-01 3116.943792     revenue
#> 24 2021-05-29 18:12:55      ETH 0.00009251 2.883485e-01 3116.943792     revenue
#> 25 2021-11-05 04:32:23     BUSD 0.10512330 1.305652e-01    1.242019     revenue
#> 26 2022-11-17 11:54:25     ETHW 0.00012050 6.168112e-04    5.118765     revenue
#> 27 2022-11-27 08:05:35     BUSD 5.77124200 7.734143e+00    1.340118         buy
#> 28 2022-11-27 08:05:35     USDC 5.77124200 7.732485e+00    1.339830        sell
#>          fees                   description comment revenue.type exchange
#> 1  1.66372374                           Buy    Spot         <NA>  binance
#> 2          NA                           Buy    Spot         <NA>  binance
#> 3  1.30299991                           Buy    Spot         <NA>  binance
#> 4          NA                           Buy    Spot         <NA>  binance
#> 5  1.21534111                           Buy    Spot         <NA>  binance
#> 6          NA                           Buy    Spot         <NA>  binance
#> 7  0.70014947                           Buy    Spot         <NA>  binance
#> 8          NA                           Buy    Spot         <NA>  binance
#> 9  0.06725739                           Buy    Spot         <NA>  binance
#> 10         NA                           Buy    Spot         <NA>  binance
#> 11 0.04708018                           Buy    Spot         <NA>  binance
#> 12         NA                           Buy    Spot         <NA>  binance
#> 13         NA             Referral Kickback    Spot      rebates  binance
#> 14         NA             Referral Kickback    Spot      rebates  binance
#> 15         NA             Referral Kickback    Spot      rebates  binance
#> 16         NA             Referral Kickback    Spot      rebates  binance
#> 17         NA             Referral Kickback    Spot      rebates  binance
#> 18         NA             Referral Kickback    Spot      rebates  binance
#> 19 6.61765817                          Sell    Spot         <NA>  binance
#> 20         NA                          Sell    Spot         <NA>  binance
#> 21 1.90881638                          Sell    Spot         <NA>  binance
#> 22         NA                          Sell    Spot         <NA>  binance
#> 23         NA             Referral Kickback    Spot      rebates  binance
#> 24         NA             Referral Kickback    Spot      rebates  binance
#> 25         NA Simple Earn Flexible Interest    Earn    interests  binance
#> 26         NA                  Distribution    Spot        forks  binance
#> 27         NA   Stablecoins Auto-Conversion    Spot         <NA>  binance
#> 28         NA   Stablecoins Auto-Conversion    Spot         <NA>  binance
#>                  rate.source
#> 1              coinmarketcap
#> 2  coinmarketcap (buy price)
#> 3              coinmarketcap
#> 4  coinmarketcap (buy price)
#> 5              coinmarketcap
#> 6  coinmarketcap (buy price)
#> 7              coinmarketcap
#> 8  coinmarketcap (buy price)
#> 9              coinmarketcap
#> 10 coinmarketcap (buy price)
#> 11             coinmarketcap
#> 12 coinmarketcap (buy price)
#> 13             coinmarketcap
#> 14             coinmarketcap
#> 15             coinmarketcap
#> 16             coinmarketcap
#> 17             coinmarketcap
#> 18             coinmarketcap
#> 19             coinmarketcap
#> 20 coinmarketcap (buy price)
#> 21             coinmarketcap
#> 22 coinmarketcap (buy price)
#> 23             coinmarketcap
#> 24             coinmarketcap
#> 25             coinmarketcap
#> 26             coinmarketcap
#> 27             coinmarketcap
#> 28             coinmarketcap
# }