Format Coinbase file

Format a .csv transaction history file from Coinbase for later ACB processing.

Usage

format_coinbase(data)

Arguments

data

The dataframe

Value

A data frame of exchange transactions, formatted for further processing.

Examples

format_coinbase(data_coinbase)
#>                  date currency    quantity total.price    spot.rate transaction
#> 1 2024-02-02 17:18:24      ETH  0.00844271       22.12 2.355250e+03     revenue
#> 2 2024-02-02 17:18:24      BTC  0.00042515       28.62 4.722199e+04     revenue
#> 3 2024-04-02 11:02:12      BTC  0.00042515       41.56 9.035295e+04        sell
#> 4 2024-04-02 11:02:12      ADA 49.57464000       41.56 8.383319e-01         buy
#> 5 2024-04-02 11:03:55      ETH  0.00844271       32.59 4.752220e+03        sell
#> 6 2024-04-02 11:03:55      ADA 40.71252100       32.59 8.004908e-01         buy
#> 7 2024-04-02 14:07:05      ADA  0.00000000        0.00 7.600000e-01        sell
#>    fees                                                       comment
#> 1  0.00 Received 0.00844271 ETH from Celsius Network LLC Crypto Di...
#> 2  0.00 Received 0.00042515 BTC from Celsius Network LLC Crypto Di...
#> 3  0.00                      Converted 0.00042515 BTC to 49.57464 ADA
#> 4 12.14                      Converted 0.00042515 BTC to 49.57464 ADA
#> 5  0.00                     Converted 0.00844271 ETH to 40.712521 ADA
#> 6  9.52                     Converted 0.00844271 ETH to 40.712521 ADA
#> 7  0.00              Sent 90.287162 ADA to abcdefghijklmnopqrstuvwxyz
#>              revenue.type exchange rate.source
#> 1 bankruptcy distribution Coinbase    exchange
#> 2 bankruptcy distribution Coinbase    exchange
#> 3                    <NA> Coinbase    exchange
#> 4                    <NA> Coinbase    exchange
#> 5                    <NA> Coinbase    exchange
#> 6                    <NA> Coinbase    exchange
#> 7                    <NA> Coinbase    exchange