Format Gemini file

Format a .csv transaction history file from Gemini for later ACB processing. Open the xlsx data file using readxl::read_excel().

Usage

format_gemini(data, list.prices = NULL, force = FALSE)

Arguments

data

The dataframe

list.prices

A list.prices object from which to fetch coin prices.

force

Whether to force recreating list.prices even though it already exists (e.g., if you added new coins or new dates).

Value

A data frame of exchange transactions, formatted for further processing.

Examples

format_gemini(data_gemini)
#> Object 'list.prices' already exists. Reusing 'list.prices'. To force a fresh download, use argument 'force = TRUE'.
#>                   date currency     quantity total.price    spot.rate
#> 1  2021-04-09 22:50:55      BTC 9.662784e-04  70.6481728 7.311369e+04
#> 2  2021-04-09 22:50:55      LTC 2.466906e-01  70.6481728 2.863837e+02
#> 3  2021-04-09 22:53:57      BTC 6.051912e-06   0.4424776 7.311369e+04
#> 4  2021-04-09 22:53:57      LTC 1.640820e-03   0.4424776 2.696686e+02
#> 5  2021-04-09 23:20:53      BAT 4.871952e+01  86.3813238 1.773033e+00
#> 6  2021-04-09 23:20:53      BTC 9.507300e-04  86.3813238 9.085789e+04
#> 7  2021-04-10 23:22:04      BTC 2.850256e-04  21.1030333 7.403909e+04
#> 8  2021-05-08 16:14:54      BAT 2.833935e+00   4.8662823 1.717147e+00
#> 9  2021-05-16 12:55:02      BAT 3.085288e+00   4.2575830 1.379963e+00
#> 10 2021-05-16 13:35:19      BAT 5.007481e+00   6.9101380 1.379963e+00
#> 11 2021-06-18 01:38:54      BAT 6.834323e+00   5.3946979 7.893537e-01
#>    transaction         fees description               comment revenue.type
#> 1          buy 2.303409e-06      LTCBTC                Market         <NA>
#> 2         sell           NA      LTCBTC                Market         <NA>
#> 3          buy 3.651810e-08      LTCBTC                 Limit         <NA>
#> 4         sell           NA      LTCBTC                 Limit         <NA>
#> 5          buy           NA      BATBTC                 Limit         <NA>
#> 6         sell 1.814241e-06      BATBTC                 Limit         <NA>
#> 7      revenue           NA      Credit Administrative Credit    referrals
#> 8      revenue           NA      Credit Administrative Credit    referrals
#> 9      revenue           NA      Credit               Deposit     airdrops
#> 10     revenue           NA      Credit               Deposit     airdrops
#> 11     revenue           NA      Credit               Deposit     airdrops
#>    exchange               rate.source
#> 1    gemini             coinmarketcap
#> 2    gemini coinmarketcap (buy price)
#> 3    gemini             coinmarketcap
#> 4    gemini coinmarketcap (buy price)
#> 5    gemini             coinmarketcap
#> 6    gemini coinmarketcap (buy price)
#> 7    gemini             coinmarketcap
#> 8    gemini             coinmarketcap
#> 9    gemini             coinmarketcap
#> 10   gemini             coinmarketcap
#> 11   gemini             coinmarketcap