Format Shakepay file

Format a .csv transaction history file from Shakepay for later ACB processing.

Usage

format_shakepay(data)

Arguments

data

The dataframe

Value

A data frame of exchange transactions, formatted for further processing.

Examples

formatted.shakepay <- format_shakepay(data_shakepay)
formatted.shakepay
#>                  date currency    quantity total.price spot.rate transaction
#> 1 2021-05-07 14:50:41      BTC  0.00103982   53.033350  51002.43         buy
#> 2 2021-05-07 21:25:36      CAD 30.00000000   30.000000      1.00     revenue
#> 3 2021-05-08 12:12:57      BTC  0.00011000    5.784024  52582.03     revenue
#> 4 2021-05-09 12:22:07      BTC  0.00012000    6.034441  50287.01     revenue
#> 5 2021-05-21 12:47:14      BTC  0.00013000    7.348590  56527.62     revenue
#> 6 2021-06-11 12:03:31      BTC  0.00014000    8.396927  59978.05     revenue
#> 7 2021-06-23 12:21:49      BTC  0.00015000    8.714877  58099.18     revenue
#> 8 2021-07-10 00:52:19      BTC  0.00052991   31.268480  59007.15        sell
#>     description  comment revenue.type exchange rate.source
#> 1 purchase/sale purchase         <NA> shakepay    exchange
#> 2         other   credit    referrals shakepay    exchange
#> 3   shakingsats   credit     airdrops shakepay    exchange
#> 4   shakingsats   credit     airdrops shakepay    exchange
#> 5   shakingsats   credit     airdrops shakepay    exchange
#> 6   shakingsats   credit     airdrops shakepay    exchange
#> 7   shakingsats   credit     airdrops shakepay    exchange
#> 8 purchase/sale     sale         <NA> shakepay    exchange