Format Uphold file

Format a .csv transaction history file from Uphold for later ACB processing.

Usage

format_uphold(data, list.prices = NULL, force = FALSE)

Arguments

data

The dataframe

list.prices

A list.prices object from which to fetch coin prices.

force

Whether to force recreating list.prices even though it already exists (e.g., if you added new coins or new dates).

Value

A data frame of exchange transactions, formatted for further processing.

Examples

format_uphold(data_uphold)
#> Object 'list.prices' already exists. Reusing 'list.prices'. To force a fresh download, use argument 'force = TRUE'.
#>                   date currency    quantity total.price   spot.rate transaction
#> 1  2021-01-07 02:40:31      BAT  1.59081275   0.5114264   0.3214875     revenue
#> 2  2021-02-09 14:26:49      BAT 12.69812163   6.8805827   0.5418583     revenue
#> 3  2021-03-06 21:32:36      BAT  0.37591275   0.3203871   0.8522910     revenue
#> 4  2021-03-07 21:46:57      LTC  0.24129740  57.0095475 236.2625850         buy
#> 5  2021-03-07 21:46:57      BAT 52.59871206  57.0095475   1.0838582        sell
#> 6  2021-03-07 21:54:09      LTC  0.00300000   0.7087878 236.2625850        sell
#> 7  2021-04-05 12:22:00      BAT  8.52198415  13.2125179   1.5504039     revenue
#> 8  2021-04-06 03:41:42      LTC  0.00300000   0.8643769 288.1256316        sell
#> 9  2021-04-06 04:47:00      LTC  0.03605981  10.3897561 288.1256316         buy
#> 10 2021-04-06 04:47:00      BAT  8.52198415  10.3897561   1.2191710        sell
#> 11 2021-05-11 07:12:24      BAT  0.47521985   0.7777261   1.6365606     revenue
#> 12 2021-06-09 04:52:23      BAT  0.67207415   0.5586404   0.8312183     revenue
#>    description         comment revenue.type exchange               rate.source
#> 1           in            <NA>     airdrops   uphold             coinmarketcap
#> 2           in            <NA>     airdrops   uphold             coinmarketcap
#> 3           in            <NA>     airdrops   uphold             coinmarketcap
#> 4        trade         BAT-LTC         <NA>   uphold             coinmarketcap
#> 5        trade         BAT-LTC         <NA>   uphold coinmarketcap (buy price)
#> 6          out withdrawal fees         <NA>   uphold             coinmarketcap
#> 7           in            <NA>     airdrops   uphold             coinmarketcap
#> 8          out withdrawal fees         <NA>   uphold             coinmarketcap
#> 9        trade         BAT-LTC         <NA>   uphold             coinmarketcap
#> 10       trade         BAT-LTC         <NA>   uphold coinmarketcap (buy price)
#> 11          in            <NA>     airdrops   uphold             coinmarketcap
#> 12          in            <NA>     airdrops   uphold             coinmarketcap