Prepare the list of coins for prices

Prepare the list of coins for prices.

Usage

prepare_list_prices(
  coins,
  start.date,
  end.date = lubridate::now("UTC"),
  currency = "CAD",
  force = FALSE
)

Arguments

coins

Which coins to include in the list.

start.date

What date to start reporting prices for.

end.date

What date to end reporting prices for.

currency

What currency to get the value of the coins in.

force

Whether to force recreating list.prices even though it already exists (e.g., if you added new coins or new dates).

Value

A data frame, with the following columns: timestamp, id, slug, name, symbol, ref_cur, open, high, low, close, volume, market_cap, time_open, time_close, time_high, time_low, spot.rate2, currency, date2.

Details

The crypto2::crypto_history API is at times a bit capricious. You might need to try a few times before it process correctly and without errors.

Examples

my.coins <- c("BTC", "ETH")
my.list.prices <- prepare_list_prices(coins = my.coins, start.date = "2023-01-01")
#> Object 'list.prices' already exists. Reusing 'list.prices'. To force a fresh download, use argument 'force = TRUE'.
head(my.list.prices)
#> # A tibble: 6 × 19
#> # Rowwise: 
#>   timestamp              id slug    name    symbol ref_cur   open   high    low
#>   <dttm>              <int> <chr>   <chr>   <chr>  <chr>    <dbl>  <dbl>  <dbl>
#> 1 2021-01-03 00:00:00     1 bitcoin Bitcoin BTC    CAD     37396. 42207. 37033.
#> 2 2021-01-04 00:00:00     1 bitcoin Bitcoin BTC    CAD     40901. 44057. 40803.
#> 3 2021-01-05 00:00:00     1 bitcoin Bitcoin BTC    CAD     41744. 42468. 36404.
#> 4 2021-01-06 00:00:00     1 bitcoin Bitcoin BTC    CAD     40889. 43642. 38539.
#> 5 2021-01-07 00:00:00     1 bitcoin Bitcoin BTC    CAD     43106. 46711. 42460.
#> 6 2021-01-08 00:00:00     1 bitcoin Bitcoin BTC    CAD     46668. 51055. 46273.
#> # ℹ 10 more variables: close <dbl>, volume <dbl>, market_cap <dbl>,
#> #   time_open <dttm>, time_close <dttm>, time_high <dttm>, time_low <dttm>,
#> #   spot.rate2 <dbl>, currency <chr>, date2 <date>