Print full crypto tax report

Will output a full crypto tax report in HTML format, which can then be printed or saved as PDF.

Usage

print_report(tax.year, name, report.info)

Arguments

tax.year

The tax year desired.

name

Name of the individual for the report.

report.info

The report info obtained from prepare_report().

Value

An HTML page containing a crypto tax report.

Examples

list.prices <- prepare_list_prices(coins = "BTC", start.date = "2021-01-01")
#> Object 'list.prices' already exists. Reusing 'list.prices'. To force a fresh download, use argument 'force = TRUE'.
all.data <- format_shakepay(data_shakepay)
formatted.ACB <- format_ACB(all.data, verbose = FALSE)
report.info <- prepare_report(formatted.ACB, list.prices)
print_report(2021, "Mr. Cryptoltruist", report.info)
#> 
#> 
#> processing file: full_report.Rmd
#>   |                                                            |                                                    |   0%  |                                                            |..                                                  |   5%                    |                                                            |.....                                               |  10% (setup)            |                                                            |.......                                             |  14%                    |                                                            |..........                                          |  19% (summary)          |                                                            |............                                        |  24%                    |                                                            |...............                                     |  29% (values)           |                                                            |.................                                   |  33%                    |                                                            |....................                                |  38% (overview)         |                                                            |......................                              |  43%                    |                                                            |.........................                           |  48% (current value)    |                                                            |...........................                         |  52%                    |                                                            |..............................                      |  57% (sup loss)         |                                                            |................................                    |  62%                    |                                                            |...................................                 |  67% (revenues)         |                                                            |.....................................               |  71%                    |                                                            |........................................            |  76% (pie_exchange)     |                                                            |..........................................          |  81%                    |                                                            |.............................................       |  86% (pie_revenue)      |                                                            |...............................................     |  90%                    |                                                            |..................................................  |  95% (unnamed-chunk-1)  |                                                            |....................................................| 100%                                                                                                                              
#> output file: full_report.knit.md
#> /usr/bin/pandoc +RTS -K512m -RTS full_report.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /home/runner/work/cryptoTax/cryptoTax/docs/reference/full_report.html --lua-filter /home/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/latex-div.lua --embed-resources --standalone --variable bs3=TRUE --section-divs --template /home/runner/work/_temp/Library/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --variable theme=bootstrap --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /tmp/RtmprdYvhD/rmarkdown-str1e9c597a8485.html 
#> 
#> Output created: full_report.html
# \dontshow{
unlink("full_report.html")
# }