Summary of gains and losses

Provides a summary of realized capital gains and losses (and total).

Usage

report_summary(
  formatted.ACB,
  today.data = TRUE,
  tax.year = "all",
  local.timezone = Sys.timezone(),
  list.prices = NULL,
  force = FALSE
)

Arguments

formatted.ACB

The formatted ACB data.

today.data

whether to fetch today's data.

tax.year

Which tax year(s) to include.

local.timezone

Which time zone to use for the date of the report.

list.prices

A list.prices object from which to fetch coin prices.

force

Whether to force recreating list.prices even though it already exists (e.g., if you added new coins or new dates).

Value

A summary data frame, containing at least the following columns: Type, Amount, currency.

Examples

all.data <- format_shakepay(data_shakepay)
formatted.ACB <- format_ACB(all.data, verbose = FALSE)
report_summary(formatted.ACB, today.data = FALSE)
#>         Type Amount currency
#> 1   tax.year    all      CAD
#> 2      gains  14.64      CAD
#> 3     losses   0.00      CAD
#> 4        net  14.64      CAD
#> 5 total.cost  36.40      CAD