Get a simple table of relevant tax information

Output a simple table with all the relevant tax information and tax form line numbers.

Usage

tax_box(report.summary, sup.losses, table.revenues, proceeds)

Arguments

report.summary

report.summary

sup.losses

sup.losses

table.revenues

table.revenues

proceeds

proceeds

Examples

my.list.prices <- prepare_list_prices(coins = "BTC", start.date = "2021-01-01")
#> Object 'list.prices' already exists. Reusing 'list.prices'. To force a fresh download, use argument 'force = TRUE'.
all.data <- format_shakepay(data_shakepay)
formatted.ACB <- format_ACB(all.data)
#> Process started at 2023-02-11 00:05:15. Please be patient as the transactions process.
#> [Formatting ACB (progress bar repeats for each coin)...]
#> Note: Adjusted cost base (ACB) and capital gains have been adjusted for the superficial loss rule. To avoid this, use argument `sup.loss = FALSE`.
#> Process ended at 2023-02-11 00:05:16. Total time elapsed: 0.01 minutes
report.summary <- report_summary(formatted.ACB, today.data = TRUE, list.prices = my.list.prices)
sup.losses <- get_sup_losses(formatted.ACB, 2021)
table.revenues <- report_revenues(formatted.ACB, 2021)
#> Note: revenues have been filtered for tax year 2021
proceeds <- get_proceeds(formatted.ACB, 2021)
#> Loss proceeds (and gains) adjusted for superficial losses. 
tax_box(report.summary, sup.losses, table.revenues, proceeds)
#>          Description    Amount
#> 1     Gains proceeds 31.268480
#> 2          Gains ACB 16.630708
#> 3              Gains 14.637772
#> 4       50% of gains  7.318886
#> 5   Outlays of gains  0.000000
#> 6    Losses proceeds  0.000000
#> 7         Losses ACB  0.000000
#> 8             Losses  0.000000
#> 9      50% of losses  0.000000
#> 10 Outlays of losses  0.000000
#> 11    Foreign income  0.000000
#> 12     Foreign gains 14.637772
#>                                                                       Comment
#> 1                                              Proceeds of sold coins (gains)
#> 2                                                   ACB of sold coins (gains)
#> 3                                                      Proceeds - ACB (gains)
#> 4                                                               Half of gains
#> 5   Expenses and trading fees (gains). Normally already integrated in the ACB
#> 6                                             Proceeds of sold coins (losses)
#> 7                                                  ACB of sold coins (losses)
#> 8                                                     Proceeds - ACB (losses)
#> 9                                                              Half of losses
#> 10 Expenses and trading fees (losses). Normally already integrated in the ACB
#> 11        Income from crypto interest or staking is considered foreign income
#> 12              Capital gains from crypto is considered foreign capital gains
#>                                             Line
#> 1                Schedule 3, line 15199 column 2
#> 2                Schedule 3, line 15199 column 3
#> 3       Schedule 3, lines 15199 column 5 & 15300
#> 4  T1, line 12700; Schedule 3, line 15300, 19900
#> 5                                   Tax software
#> 6                Schedule 3, line 15199 column 2
#> 7                Schedule 3, line 15199 column 3
#> 8       Schedule 3, lines 15199 column 5 & 15300
#> 9  T1, line 12700; Schedule 3, line 15300, 19900
#> 10                                  Tax software
#> 11                         T1, line 12100, T1135
#> 12                                         T1135